Split getImageDate into its own file to fix cyclic dependencies within serverjs/dynamo files.
In a very strange behaviour due to the circular dependencies the util.handleRouteError
function would fail because its dependency "redirect" was undefined.
Thus is a route had an Error which is caught and util.handleRouteError
called, instead of being redirected to a valid page with the error
details the server just failed.

Used dpdm (https://github.com/acrazing/dpdm) to detect the cycles and
figure out how to fix.
